requirements.txt
Flask
Flask-SQLAlchemy
psycopg[binary]
python-dotenv
simfile
beautifulsoup4
requests
